requirements.txt
fastapi
uvicorn[standard]
gunicorn
torch
torchvision
pillow
numpy
python-multipart